requirements.txt
streamlit
python-dotenv
PyPDF2
yake
transformers
langchain
langchain-google-genai
sentencepiece
